grocery_v2/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
cachelib==0.9.0
celery==5.3.6
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==23.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cssselect2==0.7.0
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
flask==3.0.0
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flask-sqlalchemy==3.1.1
fonttools==4.46.0
greenlet==3.0.2
html5lib==1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==7.0.0
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.3.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.1.3
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
Pillow==10.1.0
prompt-toolkit==3.0.41
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pydyf==0.8.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyphen==0.14.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2023.3.post1
PyYAML==5.3.1
redis==5.0.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.4
SQLAlchemy==2.0.23
ssh-import-id==5.10
systemd-python==234
tinycss2==1.2.1
Twisted==18.9.0
typing-extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vine==5.1.0
wadllib==1.3.3
wcwidth==0.2.12
weasyprint==60.1
webencodings==0.5.1
werkzeug==3.0.1
zipp==1.0.0
zope.interface==4.7.1
zopfli==0.2.3
